chore(@deip/constants): Modify  attributes constants
