+++ packages/modules/attributes/index.js
@@ -1,5 +1,4 @@
 export * from './lib/attributes-module';
-export * from './lib/constants';
 
 export { default as AttributeEdit } from './lib/components/attribute-edit';
 export { default as AttributeSet } from './lib/components/attribute-set';
