+++ packages/services/Attributes/index.js
@@ -1,2 +1,1 @@
 export * from './lib/AttributesService';
-export * from './lib/constants';
