--- packages/utilities/constants/lib/attributes/index.js
@@ -1,2 +1,0 @@
-export * from './module';
-export * from './service';
