--- packages/utilities/constants/lib/attributes/module/index.js
@@ -1,2 +1,0 @@
-export * from './rules';
-export * from './schemas';
