--- packages/utilities/constants/lib/attributes/module/rules.js
@@ -1,6 +1,0 @@
-import { ATTR_TYPES } from '../../../enum';
-
-export const ATTR_TYPES_SET_RULES = {
-  [ATTR_TYPES.TEXT]: 'required',
-  [ATTR_TYPES.TEXTAREA]: 'required'
-};
