--- packages/utilities/constants/lib/attributes/service/index.js
@@ -1,2 +1,0 @@
-export * from './types';
-export * from './scopes';
