--- packages/utilities/constants/lib/attributes/service/scopes.js
@@ -1,12 +1,0 @@
-import { ATTR_SCOPES } from '../../../enum';
-
-const ATTR_SCOPES_LABELS = {
-  [ATTR_SCOPES.PROJECT]: 'Project',
-  [ATTR_SCOPES.USER]: 'User',
-  [ATTR_SCOPES.TEAM]: 'Team'
-};
-
-export {
-  ATTR_SCOPES,
-  ATTR_SCOPES_LABELS
-};
